% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\docType{data}
\name{api_base_url}
\alias{api_base_url}
\title{Figma API infos and handlers to URLs}
\format{
An object of class \code{character} of length 1.
}
\usage{
api_base_url
}
\description{
This file contains all functions for basic
URL manipulation, and, global variables that stores
basic informations of the Figma API, like the base URL
of the API, and the implemented endpoints in this
package.
}
\keyword{datasets}
