% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{check_endpoint}
\alias{check_endpoint}
\title{Check the \code{endpoint} argument}
\usage{
check_endpoint(endpoint)
}
\arguments{
\item{endpoint}{a single string value;}
}
\description{
Check if the provided value to the \code{endpoint}
argument.
}
\details{
This function checks if the \code{endpoint} argument
is:

\itemize{
  \item{A single string value;}
  \item{One of the values in \code{figma::implemented_endpoints};}
}
}
\examples{
# Will work:
check_endpoint("files")
\dontrun{
# Will not work, because is a number
check_endpoint(1)
# Will not work, because is a vector
check_endpoint(c("a", "b", "c"))
# Will not work, because there is no "rectangle" endpoint implemented
check_endpoint("rectangle")
}
}
