% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{default_attrs}
\alias{default_attrs}
\title{Default attributes of every Figma node}
\format{
An object of class \code{character} of length 3.
}
\usage{
default_attrs
}
\description{
Every Figma document is represented as a three of nodes, and, each node have
a different type (e.g. DOCUMENT, CANVAS, TEXT, etc.). However, despite their
differences in type, every node always has three default attributes, which
are stored in the \code{figma::default_attrs} object.
}
\keyword{datasets}
