% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{document_attrs}
\alias{document_attrs}
\title{Default document attributes of a Figma file}
\format{
An object of class \code{character} of length 11.
}
\usage{
document_attrs
}
\description{
Every Figma file have some default document attributes, that is, some informations
that describe the file, like its name, the last time it was modified, etc.
These default attributes are stored in the \code{figma::document_attrs} object.
}
\keyword{datasets}
