% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{extract_canvas}
\alias{extract_canvas}
\title{Extract all canvas from a Figma Document}
\usage{
extract_canvas(document)
}
\arguments{
\item{document:}{a Figma Document (i.e. a object of class `figma_document`);}
}
\description{
Collect all canvas objects present in a Figma Document object.
}
