% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{get_endpoint_url}
\alias{get_endpoint_url}
\title{Get the URL to each of the implemented endpoints of Figma API}
\usage{
get_endpoint_url(endpoint)
}
\arguments{
\item{endpoint}{a single string with the name of the desired endpoint;}
}
\value{
a string with the URL to the given endpoint.
}
\description{
Get the URL to each of the implemented endpoints of Figma API
}
\details{
This function accepts only single string values,
and, this string needs to be one of the values present
in \code{figma::implemented_endpoints}.

If the user provided a NULL or NA value,
or, a vector of strings, the function will
automatically prompt the user with an error.
}
\examples{
# Returns the URL to the `files` endpoint of Figma API
get_endpoint_url("files")

}
