% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma.R
\name{get_figma_file}
\alias{get_figma_file}
\title{Get a Figma File from the API}
\usage{
get_figma_file(file_key, token, .output_format = "response")
}
\arguments{
\item{file_key}{A string with the key of the Figma File you want to get;}

\item{token}{A string with your personal Figma token to authenticate in the API;}

\item{.output_format}{The output format (i.e. you want your Figma data as is?
Or, you want to allocate it on a `tibble`? etc...)}
}
\value{
A Figma Document object (i.e. a object of class `figma_document`);
}
\description{
Use the \code{/v1/files/} endpoint of Figma API to get a Figma File as an R object.
}
\details{
You can easily get the key of your Figma File from the URL of this file.
}
