% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma.R
\name{get_figma_objects}
\alias{get_figma_objects}
\title{Get data of a set of objects in a Figma File from the API}
\usage{
get_figma_objects(
  file_key,
  token,
  node_id,
  geometry = FALSE,
  .output_format = "response",
  ...
)
}
\arguments{
\item{file_key}{A string with the key of the Figma File you want to get;}

\item{token}{A string with your personal Figma token to authenticate in the API;}

\item{node_id}{A string with the node ID (or a vector of strings with node IDs);}

\item{geometry}{A boolean value indicating if you want to export vector data.
Defaults to FALSE;}

\item{.output_format}{The output format. Options are \code{"response",
"figma_document", "tibble"}. Defaults to \code{"response"};}

\item{...}{Further arguments that are passed to \code{parse_response_object()};}
}
\description{
This function uses the \code{/v1/files/} endpoint of Figma API
to get the data of an specific object (or a set of objects) from a Figma file,
and fit it into a R object.
}
\details{
Every object draw in a canvas/page of a Figma file, is represented as a node in
the canvas on which it was drawn. Because of this notion, to identify the specific
object (or "node") that you want to search, you need to give the ID code that identifies
this specific object in the canvas.

You can easily get this ID from the URL that appears in your browser when
you select this specific object. See vignette.
}
