% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\docType{data}
\name{standard_attrs}
\alias{standard_attrs}
\title{Parse the response object returned by Figma API}
\format{
An object of class \code{character} of length 3.
}
\usage{
standard_attrs
}
\description{
Parse the response object returned by Figma API
}
\keyword{datasets}
