% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{as_figma_document}
\alias{as_figma_document}
\title{Convert a `httr` response object to a Figma Document object}
\usage{
as_figma_document(response)
}
\arguments{
\item{response:}{a `response` object produced by a `httr` HTTP method (e.g. `httr::GET()`, `httr::POST()`, etc.);}
}
\description{
This function is not intended for client/user-use.
}
