% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{build_query_string}
\alias{build_query_string}
\title{Build a query string from a set of named parameters}
\usage{
build_query_string(parameters)
}
\arguments{
\item{parameters}{A list with a set of key-value pairs to compose the query string}
}
\value{
A single string with the query string produced.
}
\description{
Utility function used to build query strings.
}
\details{
This function takes a set of key-value pairs (or in other words,
a set of named arguments), to build a query string. It basically
combine (or "collapse") all key-value pairs together, to form
the resulting query string.

Logical values (TRUE or FALSE) are automatically converted to
a lower-case version ("true" or "false"), since these versions are
more tipically used in standard query strings.
}
\examples{
build_query_string(list(skip = 100, page = 2, title = "Document"))
# Returns a empty string:
build_query_string(list())
build_query_string(list(a = 1, b = 2, c = TRUE))

}
