% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma.R
\name{get_figma_file}
\alias{get_figma_file}
\title{Get a Figma File from the API}
\usage{
get_figma_file(file_key, token)
}
\arguments{
\item{file_key}{A string with the key of the Figma File you want to get;}

\item{token}{A string with your personal Figma token to authenticate in the API;}
}
\value{
A Figma Document object (i.e. a object of class `figma_document`);
}
\description{
This function will use the \code{/v1/files/} endpoint to get a Figma File
from the Figma API.
}
\details{
You can easily get the key of your Figma File from the URL of this file.
In other words, when you visit this file in your web browser, the URL
will probably look something like this:

\code{https://www.figma.com/file/hch8YlkIrYbU3raDzjPvCz/Untitled}

In the above example, the string \code{hch8YlkIrYbU3raDzjPvCz} is the
key to the file. So, every URL to a Figma File comes in the following format:

\code{https://www.figma.com/file/file_key/file_title}
}
